更新

+++ delete.py
@@ -4,7 +4,7 @@ def deleteMp4(folderPath):
     files = os.listdir(folderPath)
     originFile = folderPath.split(os.path.sep)[-1] + '.mp4'
     for file in files:
-        if file != originFile:
+        if file != originFile and file.endswith('.mp4'):
             os.remove(os.path.join(folderPath, file))
